// src/file-parser/student.dto.ts
export class StudentDTO {
  id: number;
  studentCode: string;
  cauCode: string;
  cauSegment: string;
  cauName: string;
  modality: string;
  shift: string;
  level: string;
  riPeriod: string;
  listPriceMm: number;
  listPriceAb: number;
  diffListPriceGrade: number;
  currentPromMm: number;
  ticketAProm: number;
  ticketBProm: number;
  ticketDProm: number;
  listPrice1Course: number;
  listPrice3Course: number;
  listPrice4Course: number;
  listPrice6Course: number;
  additionalProm: number;
  pFeeProm: number;
  mmTotal: number;
  ticketATotal: number;
  ticketBTotal: number;
  ticketDTotal: number;
}